+++ server/gameserver/realm/entity/player/Player.Variables.cs
@@ -59,13 +59,13 @@ namespace LoESoft.GameServer.realm.entity.player
         public Client Client { get; }
         public int Credits { get; set; }
         public int Tokens { get; set; }
-        public int CurrentFame { get; set; }
-        public int Experience { get; set; }
-        public int ExperienceGoal { get; set; }
-        public int Fame { get; set; }
+        public double CurrentFame { get; set; }
+        public double Experience { get; set; }
+        public double ExperienceGoal { get; set; }
+        public double Fame { get; set; }
         public FameCounter FameCounter { get; }
         public TaskManager TaskManager { get; }
-        public int FameGoal { get; set; }
+        public double FameGoal { get; set; }
         public bool Glowing { get; set; }
         public bool HasBackpack { get; set; }
         public int HealthPotions { get; set; }
